api/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
mysql-connector-python==8.4.0